VIEWS

+++ src/main/java/com/jewelryguard/controller/JewelryController.java
@@ -3,6 +3,11 @@ package com.jewelryguard.controller;
 import com.jewelryguard.model.*;
 import com.jewelryguard.service.*;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
@@ -34,23 +39,8 @@ public class JewelryController {
 	@Autowired
 	private MyFileService myFileService;
 
-	@RequestMapping(value="/new", method = RequestMethod.GET)
-	public ModelAndView create(){
-		ModelAndView modelAndView = new ModelAndView();
-		Jewelry jewelry = new Jewelry();
-		List<Category> categoryList = categoryService.findAll();
-		modelAndView.addObject("categoryList", categoryList);
 
-		List<Metal> metalList = metalService.findAll();
-		modelAndView.addObject("metalList", metalList);
 
-		List<MyFile> fileList = myFileService.findAll();
-		modelAndView.addObject("files", fileList);
-
-		modelAndView.addObject("jewelry", jewelry);
-		modelAndView.setViewName("jewelrys/new");
-		return modelAndView;
-	}
 
 	/*
 	 * Multiple
@@ -73,20 +63,29 @@ public class JewelryController {
 			modelAndView.addObject("successMessage", "Dodano do biblioteki.");
 			modelAndView.addObject("jewelryList", jewelryList);
 
-			modelAndView.setViewName("jewelrys");
+			modelAndView.setViewName("jewelrys/list");
 
 		}
 		return modelAndView;
 	}
 
 	@RequestMapping(method = RequestMethod.GET)
-	public ModelAndView showAll(){
+	public ModelAndView showAll(@RequestParam(value = "page", required = false) Integer page, @RequestParam(value = "limit", required = false)Integer limit){
 		ModelAndView modelAndView = new ModelAndView();
 		Authentication auth = SecurityContextHolder.getContext().getAuthentication();
 		User user = userService.findUserByEmail(auth.getName());
-		List<Jewelry> jewelryList = user.getJewelryList();
+		if(page == null){
+			page = 0;
+		}
+		if(limit == null){
+			limit = 10;
+		}
+		Pageable pageable = createPageRequest(page,limit);
+		Page<Jewelry> jewelryList = jewelryService.findAllByUser(user, pageable);
+
+
 		modelAndView.addObject("jewelryList", jewelryList );
-		modelAndView.setViewName("jewelrys");
+		modelAndView.setViewName("jewelrys/list");
 		return modelAndView;
 	}
 
@@ -95,7 +94,23 @@ public class JewelryController {
 	 *
 	 *
 	 */
+    @RequestMapping(value="/new", method = RequestMethod.GET)
+    public ModelAndView create(){
+        ModelAndView modelAndView = new ModelAndView();
+        Jewelry jewelry = new Jewelry();
+        List<Category> categoryList = categoryService.findAll();
+        modelAndView.addObject("categoryList", categoryList);
 
+        List<Metal> metalList = metalService.findAll();
+        modelAndView.addObject("metalList", metalList);
+
+        List<MyFile> fileList = myFileService.findAll();
+        modelAndView.addObject("files", fileList);
+
+        modelAndView.addObject("jewelry", jewelry);
+        modelAndView.setViewName("jewelrys/new");
+        return modelAndView;
+    }
 	@RequestMapping(value="/{jewelryId}", method = RequestMethod.GET)
 	public ModelAndView showOne(@PathVariable("jewelryId") int jewelryId){
 		ModelAndView modelAndView = new ModelAndView();
@@ -108,9 +123,11 @@ public class JewelryController {
 				stream().
 				map(MyFile::getId).
 				collect(Collectors.toList());
-
+		MyFile myFile = new MyFile();
+		myFile.setJewelry(jewelry);
 		modelAndView.addObject("jewelry", jewelry);
 		modelAndView.addObject("images", images);
+		modelAndView.addObject("myFile", myFile);
 		modelAndView.setViewName("jewelrys/index");
 		return modelAndView;
 	}
@@ -130,16 +147,25 @@ public class JewelryController {
 	}
 
 	@RequestMapping(value="/{jewelryId}", method = RequestMethod.DELETE)
-	public ModelAndView deleteOne(@PathVariable("jewelryId") int jewelryId){
+	public ModelAndView deleteOne( @PathVariable("jewelryId") int jewelryId){
 		ModelAndView modelAndView = new ModelAndView();
+
 		Authentication auth = SecurityContextHolder.getContext().getAuthentication();
 		User user = userService.findUserByEmail(auth.getName());
+
 		Jewelry jewelry = jewelryService.findOne(jewelryId);
+		if ( !jewelry.getUser().getEmail().equals(auth.getName()) ) {
+			modelAndView.setViewName("jguard");
+			throw new AccessDeniedException("You have no access!");
+		}
+
 		boolean isDeleted = jewelryService.delete(jewelry);
-		List<Jewelry> jewelryList = jewelryService.findAllByUser(user);
+		Page<Jewelry> jewelryList = jewelryService.findAllByUser(user, createPageRequest(0,10));
+
 		modelAndView.addObject("message", isDeleted ? "Success: deleted jewelry.":"Error");
 		modelAndView.addObject("jewelryList", jewelryList);
-		modelAndView.setViewName("jewelrys");
+		modelAndView.setViewName("jewelrys/list");
+
 		return modelAndView;
 	}
 
@@ -163,27 +189,54 @@ public class JewelryController {
 		} catch (FileNotFoundException e) {
 			e.printStackTrace();
 		}
-//		HashMap<String, String> images = myFileService.findAllByJewelry(jewelry);
-		modelAndView.addObject("images", myFileService.findAll().stream().map(MyFile::getPath).collect(Collectors.toList()));
-		modelAndView.addObject("myFile", new MyFile() );
-		modelAndView.setViewName("jewelrys/index");
-		return modelAndView;
-	}
-
-	@RequestMapping(value="{jewelryId}/images", method = RequestMethod.GET)
-	public ModelAndView showImages(@PathVariable("jewelryId") int jewelryId) {
-		ModelAndView modelAndView = new ModelAndView();
-		Jewelry jewelry = jewelryService.findOne(jewelryId);
-		if (jewelry == null) {
-			throw new EntityNotFoundException("Not found jewewlry for id: " + jewelryId);
-		}
-
-//		HashMap<String, String> images = myFileService.findAllByJewelry(jewelry);
+		modelAndView.addObject( "jewelry", jewelry);
 		modelAndView.addObject("images", myFileService.findAll().stream().map(MyFile::getId).collect(Collectors.toList()));
-		modelAndView.addObject("myFile", new MyFile());
-		modelAndView.setViewName("jewelrys/images");
+
+		modelAndView.setViewName("comp/j-img-list");
 		return modelAndView;
 	}
 
+	private Pageable createPageRequest(Integer page, Integer limit){
+		return new Pageable() {
+            @Override
+            public int getPageNumber() {
+                return 0;
+            }
 
+            @Override
+            public int getPageSize() {
+                return 0;
+            }
+
+            @Override
+            public int getOffset() {
+                return 0;
+            }
+
+            @Override
+            public Sort getSort() {
+                return null;
+            }
+
+            @Override
+            public Pageable next() {
+                return null;
+            }
+
+            @Override
+            public Pageable previousOrFirst() {
+                return null;
+            }
+
+            @Override
+            public Pageable first() {
+                return null;
+            }
+
+            @Override
+            public boolean hasPrevious() {
+                return false;
+            }
+        };
+	}
 }
